requirements.txt
# SPT Server Editor - Зависимости проекта
# Основные библиотеки для работы с GUI и данными

# HTTP запросы для сканирования базы данных
requests>=2.31.0

# Типизация (для Python < 3.9)
typing-extensions>=4.0.0

# Улучшение производительности
orjson>=3.9.0          # Ускорение JSON операций
httpx>=0.24.0          # Улучшение HTTP запросов
pyyaml>=6.0            # Работа с конфигурационными файлами

# Встроенные модули Python (не требуют установки):
# - tkinter (GUI Framework)
# - json (работа с JSON) - заменен на orjson
# - pathlib (работа с путями)
# - logging (логирование)
# - datetime (даты и время)
# - os, sys, subprocess (системные операции)
# - time (время)
# - shutil (файловые операции)

# Для создания исполняемых файлов (опционально)
# pyinstaller>=5.0

# Для тестирования (опционально)
# pytest>=7.0.0
# pytest-cov>=4.0.0

# Для линтинга кода (опционально)
# flake8>=6.0.0
# black>=23.0.0
